+++ pkg/ui/react-app/src/utils/index.ts
@@ -287,7 +287,8 @@ export const encodePanelOptionsToQueryString = (panels: PanelMeta[]): string => 
 };
 
 export const createExpressionLink = (expr: string): string => {
-  return `../graph?g0.expr=${encodeURIComponent(expr)}&g0.tab=1&g0.stacked=0&g0.range_input=1h`;
+  const alertSourceTemplate = '/graph?g0.expr={{.Expr}}&g0.tab=1';
+  return `${alertSourceTemplate.replace('{{.Expr}}', encodeURIComponent(expr))}&g0.stacked=0&g0.range_input=1h`;
 };
 
 export const createExternalExpressionLink = (expr: string): string => {